# hero/openapi.json
{
  "openapi" : "3.1.0",
  "components" : {
    "schemas" : {
      "Hero" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string",
            "minLength" : 3,
            "maxLength" : 50
          },
          "otherName" : {
            "type" : "string"
          },
          "level" : {
            "type" : "integer",
            "format" : "int32",
            "minimum" : 1
          },
          "picture" : {
            "type" : "string"
          },
          "powers" : {
            "type" : "string"
          }
        },
        "required" : [ "name", "level" ]
      }
    }
  },
  "paths" : {
    "/api/heroes" : {
      "put" : {
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Hero"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Hero"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Update Hero",
        "tags" : [ "Hero Resource" ]
      },
      "get" : {
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Hero"
                  }
                }
              }
            }
          }
        },
        "summary" : "Get All Heroes",
        "tags" : [ "Hero Resource" ]
      },
      "post" : {
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Hero"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "string",
                  "format" : "uri"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Create Hero",
        "tags" : [ "Hero Resource" ]
      }
    },
    "/api/heroes/hello" : {
      "get" : {
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "summary" : "Hello",
        "tags" : [ "Hero Resource" ]
      }
    },
    "/api/heroes/random" : {
      "get" : {
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Hero"
                }
              }
            }
          }
        },
        "summary" : "Get Random Hero",
        "tags" : [ "Hero Resource" ]
      }
    },
    "/api/heroes/{id}" : {
      "get" : {
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Hero"
                }
              }
            }
          }
        },
        "summary" : "Get Hero",
        "tags" : [ "Hero Resource" ]
      },
      "delete" : {
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "summary" : "Delete Hero",
        "tags" : [ "Hero Resource" ]
      }
    }
  },
  "info" : {
    "title" : "hero-service API",
    "version" : "1.0.0-SNAPSHOT"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ]
}